third_party/WebKit/LayoutTests/webexposed/event-target-in-prototype.html

<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
description("Check that interfaces inherit from EventTarget");

eventTargets = ['EventSource', 'FileReader', 'MediaController', 'MediaSource', 'MessagePort', 'Node', 'Performance', 'SVGElementInstance', 'SharedWorker', 'WebSocket', 'Window', 'Worker', 'AudioContext'];

for (var i = 0; i < eventTargets.length; i++) {
    shouldBe(eventTargets[i] + ".prototype.__proto__", "EventTarget.prototype");
}

</script>
</body>
</html>
